requirements.txt
ipykernel
openai
python-dotenv
yfinance
pandas
matplotlibpip